Properly cleaned up the Translation Builder code. Made the rendering of the bounding box for word finding correct again. Now just need to properly make the translation work...

+++ MassDictionary/Translation Page/TranslationBox.swift
@@ -14,10 +14,10 @@ enum WordType {
 // A translation box with the x, y as the top left corner (?)
 struct TranslationBox: Identifiable {
     let id = UUID()
-    let x: Int
-    let y: Int
-    let width: Int
-    let height: Int
+    let x: CGFloat
+    let y: CGFloat
+    let width: CGFloat
+    let height: CGFloat
     let type: WordType
     let translation: String
